requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
groq>=0.4.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
pandas>=2.0.0
openpyxl>=3.1.0
Pillow>=10.0.0
django-cors-headers>=4.3.0
